Be able to resolve some of the easier examples
When everything is nice and clear, it succeeds to fill the board.
Complicated stuff <=> empty tiles not touching numbered tiles.
Simple pointing, using "hidden singles" should fix most of the stuff.

Code needs some polishing, it was simplified alot over the time, but
there is really no difference between TILE_{1,2,3}, just the number
in the condition is not the same.